Update Python version in build.yml to 3.12
Python 3.7 has fallen out of support and could not install the latest
versions of build tools like setuptools and cibuildwheel.

Also, the setup-python action was updated to v5 as the v1 version was
producing warnings about end of life.

With the latest version of setuptools, `python setup.py sdist` no longer
works so the sdist creation step was modified to use pypa/build instead.